Start refactoring files structure.
