--- src/components/PageContainer.jsx
@@ -1,11 +1,0 @@
-// components/PageContainer.js
-import styled from 'styled-components';
-
-const PageContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  position: relative;
-`;
-
-export default PageContainer;
