+++ src/routes.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { HashRouter, Routes, Route } from 'react-router-dom';
-import Welcome from './screens/Welcome';
-import Community from './screens/Community';
-import AskQuestionPage from './components/AskQuestionPage';
+import WelcomePage from './screens/WelcomePage';
+import CommunityPage from './screens/CommunityPage';
+import AskQuestionPage from './screens/AskQuestionPage';
 import QuestionDetail from './components/QuestionDetail';
 
 const RouterPage = () => {
   return (
     <HashRouter>
       <Routes>
-        <Route path="/" element={<Welcome />} />
-        <Route path="/community/*" element={<Community />} />
+        <Route path="/" element={<WelcomePage />} />
+        <Route path="/community/*" element={<CommunityPage />} />
         <Route path="/askquestion/:did" element={<AskQuestionPage />} />
         <Route path="/question/:did" element={<QuestionDetail />} />
       </Routes>
