+++ src/styled-components/FormHeader.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const FormHeader = styled.h2`
+  color: #FFFF00;
+  margin: 0;
+
+  @media (max-width: 1000px) {
+    font-size: 14px;
+  }
+`;
+
+export default FormHeader;
