+++ src/styled-components/FormItem.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const FormItem = styled.div`
+  margin-bottom: 15px;
+
+  @media (max-width: 1000px) {
+    margin-bottom: 10px;
+  }
+`;
+
+export default FormItem;
