+++ src/styled-components/index.js
@@ -1,0 +1,9 @@
+export { default as Button } from './Button';
+export { default as CodeEditor } from './CodeEditor';
+export { default as FormHeader } from './FormHeader';
+export { default as FormHeaderWrapper } from './FormHeaderWrapper';
+export { default as FormItem } from './FormItem';
+export { default as FormWrapper } from './FormWrapper';
+export { default as Input } from './Input';
+export { default as PageContainer } from './PageContainer';
+export { default as TextArea } from './TextArea';
